changed speed to float
for some reason its stopp at boundary

+++ include/obj_texture.h
@@ -42,7 +42,7 @@ public:
   int mPitch = 0;
 
   //Debris specific stuff
-  int mSpeed = 0;
+  float mSpeed = 0;
   vect2 mDirection;
